Camera Works properly now! Force2D now has some vector functions

+++ MacGyver/Vector3.h
@@ -6,7 +6,7 @@ namespace Macgyver::Math{
 	class Vector3 {
 	public:
 		float x, y, z;
-		Vector3(int x = 0, int y = 0, int z = 0);
+		Vector3(float x = 0, float y = 0, float z = 0);
 		Vector3 operator+(const Vector3& a);
 		Vector3 operator-(const Vector3& a);
 		Vector3 operator*(const float b);
